coprocessor: move build related function to DAGBuilder
