+++ src/coprocessor/dag/mod.rs
@@ -32,5 +32,6 @@ pub mod rpn_expr;
 mod scanner;
 
 pub use self::batch_handler::BatchDAGHandler;
+pub use self::builder::DAGBuilder;
 pub use self::handler::DAGRequestHandler;
 pub use self::scanner::{ScanOn, Scanner};
